temporal/pyproject.toml
[project]
name = "agritech-worker"
version = "0.1.0"
description = ""
authors = [
    {name = "mayank138008",email = "mayank.pal1909@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio (>=1.15.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "numpy (==1.23.5)",
    "rasterio (==1.3.9)",
    "geopandas (==0.12.2)",
    "joblib (>=1.5.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "scikit-learn (==1.2.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
